+++ test/delegator_test.exs
@@ -48,7 +48,7 @@ defmodule DelegatorTest do
       @moduledoc false
       import Delegator
 
-      defdelegateall A
+      defdelegateall to: A
     end
 
     test "delegates all functions in A" do
@@ -64,7 +64,9 @@ defmodule DelegatorTest do
 
     defmodule DelegatedFunctionsAreOverridable do
       @moduledoc false
-      use Delegator, to: A
+      import Delegator
+
+      defdelegateall to: A
 
       def a(_, _, _), do: 5
     end
@@ -79,11 +81,11 @@ defmodule DelegatorTest do
       @moduledoc false
       import Delegator
 
-      defdelegateall A, as: [a: :c]
-      defdelegateall A, except: [a: 0]
-      defdelegateall B, only: [b: 0]
-      defdelegateall B, prefix: :before
-      defdelegateall B, suffix: :after
+      defdelegateall to: A, as: [a: :c]
+      defdelegateall to: A, except: [a: 0]
+      defdelegateall to: B, only: [b: 0]
+      defdelegateall to: B, prefix: :before
+      defdelegateall to: B, suffix: :after
     end
 
     test "aliases all a/* functions in A as c/*" do
@@ -150,7 +152,7 @@ defmodule DelegatorTest do
       @moduledoc false
       import Delegator
 
-      defdelegateallmacros A
+      defdelegateallmacros to: A
     end
 
     test "delegates all macros in A" do
@@ -173,11 +175,11 @@ defmodule DelegatorTest do
       @moduledoc false
       import Delegator
 
-      defdelegateallmacros A, as: [m: :o]
-      defdelegateallmacros A, except: [m: 1]
-      defdelegateallmacros B, only: [n: 1]
-      defdelegateallmacros B, prefix: :before
-      defdelegateallmacros B, suffix: :after
+      defdelegateallmacros to: A, as: [m: :o]
+      defdelegateallmacros to: A, except: [m: 1]
+      defdelegateallmacros to: B, only: [n: 1]
+      defdelegateallmacros to: B, prefix: :before
+      defdelegateallmacros to: B, suffix: :after
     end
 
     test "aliases all m/* macros in A as o/*" do
@@ -226,7 +228,7 @@ defmodule DelegatorTest do
       @moduledoc false
       import Delegator
 
-      defdelegateeverything A
+      defdelegateeverything to: A
     end
 
     test "delegates all functions in A" do
@@ -260,9 +262,9 @@ defmodule DelegatorTest do
       @moduledoc false
       import Delegator
 
-      defdelegateeverything A, as: [a: :c, m: :o]
-      defdelegateeverything A, except: [a: 0, m: 1]
-      defdelegateeverything B, only: [b: 0, n: 1]
+      defdelegateeverything to: A, as: [a: :c, m: :o]
+      defdelegateeverything to: A, except: [a: 0, m: 1]
+      defdelegateeverything to: B, only: [b: 0, n: 1]
     end
 
     test "aliases all a/* functions in A as c/*" do
@@ -309,57 +311,6 @@ defmodule DelegatorTest do
       assert DefDelegateEverythingWithOpts.n(1) == [1]
       refute_macro DefDelegateEverythingWithOpts, :n, 2
       refute_macro DefDelegateEverythingWithOpts, :n, 3
-    end
-  end
-
-  describe "use Delegator to a single module" do
-    defmodule DelegateEverythingToOne do
-      @moduledoc false
-      use Delegator, to: A
-    end
-
-    test "delegates all functions in A" do
-      assert DelegateEverythingToOne.a() == 1
-      assert DelegateEverythingToOne.a(1) == 2
-      assert DelegateEverythingToOne.a(1, 2) == 3
-      assert DelegateEverythingToOne.a(1, 2, 3) == 4
-    end
-
-    test "delegates all macros in A" do
-      require DelegateEverythingToOne
-
-      assert DelegateEverythingToOne.m(1) == [1]
-      assert DelegateEverythingToOne.m(1, 2) == [1, 2]
-      assert DelegateEverythingToOne.m(1, 2, 3) == [1, 2, 3]
-    end
-  end
-
-  describe "use Delegator to multiple modules" do
-    defmodule DelegateEverythingToMany do
-      @moduledoc false
-      use Delegator, to: [A, B]
-    end
-
-    test "delegates all functions" do
-      assert DelegateEverythingToMany.a() == 1
-      assert DelegateEverythingToMany.a(1) == 2
-      assert DelegateEverythingToMany.a(1, 2) == 3
-      assert DelegateEverythingToMany.a(1, 2, 3) == 4
-      assert DelegateEverythingToMany.b() == 1
-      assert DelegateEverythingToMany.b(1) == 2
-      assert DelegateEverythingToMany.b(1, 2) == 3
-      assert DelegateEverythingToMany.b(1, 2, 3) == 4
-    end
-
-    test "delegates all macros" do
-      require DelegateEverythingToMany
-
-      assert DelegateEverythingToMany.m(1) == [1]
-      assert DelegateEverythingToMany.m(1, 2) == [1, 2]
-      assert DelegateEverythingToMany.m(1, 2, 3) == [1, 2, 3]
-      assert DelegateEverythingToMany.n(1) == [1]
-      assert DelegateEverythingToMany.n(1, 2) == [2, 1]
-      assert DelegateEverythingToMany.n(1, 2, 3) == [3, 2, 1]
     end
   end
 end